chore: removed MediaPlayerContext
