--- app/contexts/index.ts
@@ -1,4 +1,0 @@
-export {
-  MediaPlayerContext,
-  MediaPlayerProvider,
-} from './media-player.context';
